Removed the counterDisplayHandleGesture into it's own utility code, and now CounterDiaplay component calls it when needed, leaving the CounterDisplay just withthe shared values for prop drilling, and one useState to change the color of the display of the component itself
