+++ components/Counter/CounterDisplay.jsx
@@ -10,8 +10,9 @@ import Animated, {
 } from 'react-native-reanimated'
 import { animatedMove } from '../../utils/animatedStyles'
 import { responsivePixel } from '../../utils/responsivePixel'
+import { counterDisplayHandleGesture } from '../../utils/counterDisplayHandleGesture'
 
-const CounterDisplay = ({ id, counterName, func, count }) => {
+const CounterDisplay = ({ id, counterName, setCount, count }) => {
   // Center of the Counter
   const startingPosition = 0
   // State for reactive coloring
@@ -26,80 +27,8 @@ const CounterDisplay = ({ id, counterName, func, count }) => {
   // Animated styles for changing the position of the display
   const animatedStyles = animatedMove(x, y)
 
-  // Functions that increment, decrement, reset and remove the counter
-  const add = () => {
-    Vibration.vibrate(10)
-    const newCount = count + 1
-
-    AsyncStorage.setItem(id, JSON.stringify({ "counterName": counterName, "count": newCount }))
-      .then(func(newCount))
-      .catch(err => console.log(err))
-  }
-  const less = () => {
-    Vibration.vibrate(10)
-    const newCount = count < 1 ? 0 : count - 1
-    AsyncStorage.setItem(id, JSON.stringify({ "counterName": counterName, "count": newCount }))
-      .then(func(newCount))
-      .catch(err => console.log(err))
-  }
-  const reset = () => {
-    AsyncStorage.setItem(id, JSON.stringify({ "counterName": counterName, "count": 0 }))
-      .then(func(0))
-      .catch(err => console.log(err))
-  }
-  const remove = () => {
-    Vibration.vibrate(10)
-    AsyncStorage.removeItem(id)
-      .then(() => func(null))
-      .catch(err => console.log(err))
-  }
   // Event hanlder provided by Reanimated that is specifically designed to work with the gesture-handler package
-  const eventHandler = useAnimatedGestureHandler({
-    // Starting position of the display
-    onStart: (event, ctx) => {
-
-    },
-    onActive: (event, ctx) => {
-      // This change thr actual position of the display
-      x.value = event.translationX
-      y.value = event.translationY
-      // Limiting movement
-      y.value <= 0 ? y.value = 0
-        : y.value >= 65 ? y.value = 65 : y.value
-      x.value >= 60 ? x.value = 60
-        : x.value <= -60 ? x.value = -60 : x.value
-      // Executing methods, and changing the previus states by position
-
-      if (y.value >= 65 && x.value <= 50 && x.value >= -50) runOnJS(setReseting)(true)
-      else runOnJS(setReseting)(false)
-
-    },
-    onEnd: (event, ctx) => {
-      // Verifies the last position when the display was released
-      // in order to execute the corresponding methods and states
-      if (x.value > 0) runOnJS(add)()
-      else if (x.value < 0) runOnJS(less)()
-
-      if (y.value <= -65 && x.value <= 50 && x.value >= -50) runOnJS(remove)()
-      else if (y.value >= 65 && x.value <= 50 && x.value >= -50) runOnJS(reset)()
-
-      runOnJS(setReseting)(false)
-      // Restarting the display position
-      x.value = withSpring(startingPosition, {
-        damping: 10,
-        mass: .5,
-        stiffness: 300,
-        overshootClamping: false,
-      })
-      y.value = withSpring(startingPosition, {
-        damping: 10,
-        mass: .5,
-        stiffness: 300,
-        overshootClamping: false,
-      })
-    }
-  })
-
+  const eventHandler =  counterDisplayHandleGesture(x, y, startingPosition, id, counterName, setCount, count, setReseting)
   return <View>
     <PanGestureHandler onGestureEvent={eventHandler}>
       <Animated.View style={[animatedStyles, styles.animatedContainer]}>
